apiclient: apply: consolidate cloud resolvers
Move AWS-specific resolvers (S3, Secrets Manager, SSM) into a separate
cloud_resolvers module with feature-gated CloudError enum. Improve
error handling, documentation, and test coverage.
