+++ sources/api/apiclient/src/lib.rs
@@ -29,6 +29,8 @@ pub mod reboot;
 pub mod report;
 pub mod set;
 pub mod update;
+#[cfg(feature = "tls")]
+pub mod cloud_resolvers;
 pub mod uri_resolver;
 
 mod error {
